pyproject.toml
[project]
name = "pleiades"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["httpx>=0.27.2", "pydantic>=2.8.2", "python-dotenv>=1.0.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["examples"]
